+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+## [1.0.0] (2024-04-05)
+
+> [!WARNING]  
+> Potentially BREAKING release
+
+- feat!: update icons to now include 1898 Carbon icons
+- feat: add SVG cleaning step to generator script
+- feat: consolidate generator script into one Dart file
+- docs: update documentation
+
 ## [0.0.1+2]
 
-* Fix image preview in README.
+- Fix image preview in README.
 
 ## [0.0.1]
 
-* Initial release.
+- Initial release.